requirements.txt
langchain-core>=0.1.30
langchain-groq>=0.1.5
langgraph>=0.0.34
pydantic>=2.6.4
python-dotenv>=1.0.1
typing-extensions>=4.10.0

# Optional but recommended for async support and clean I/O
aiohttp>=3.9.5

